% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDataType.R
\name{convertDataType}
\alias{convertDataType}
\title{convertDataType}
\usage{
convertDataType(x)
}
\arguments{
\item{x}{list of char. variables}
}
\value{
list of variables with tree types(character,numeric,date)
}
\description{
Get a list of char. variables and convert them into character,
numeric, and date types. The function finds the first, not NA element,
checks what is the type of that element (char, num, date),
and converts the whole list into this type.
}
