% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameFromXML.R
\name{dataFrameFromXML}
\alias{dataFrameFromXML}
\title{dataFrameFromXML}
\usage{
dataFrameFromXML(xmlFilePath)
}
\arguments{
\item{xmlFilePath}{Path to .xml file}
}
\value{
data.frame or warning in case of missing file
}
\description{
Function creates data.frame from .xml file.
It analyzes the content of every column and converts it into one of the types:
character, numeric, date.
}
