% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_not_NA.R
\name{firstNotNA}
\alias{firstNotNA}
\title{First not NA element in a set}
\usage{
firstNotNA(x)
}
\arguments{
\item{x}{List.}
}
\value{
single value
}
\description{
This function goes through the list of elements as long as NA occurs.
If the first element from the list is not an NA It won't go further.
}
