% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDate.R
\name{isDate}
\alias{isDate}
\title{isDate}
\usage{
isDate(x)
}
\arguments{
\item{x}{(char. type)}
}
\value{
boolen value TRUE/FALSE
}
\description{
This function checks if a given variable is representing some date.
}
\examples{
isDate("ASD")
isDate("2016-01-12")
}
