% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numeric.R
\name{isNumeric}
\alias{isNumeric}
\title{isNumeric}
\usage{
isNumeric(x)
}
\arguments{
\item{x}{A number/char.}
}
\value{
boolen value TRUE/FALSE
}
\description{
This function checks if a given variable is a numeric value.
If the character type variable contains only numbers function will return true.
}
\examples{
isNumeric("ASD")
isNumeric("213")
}
